requirements.txt
selenium==4.17.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
pandas==2.2.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
undetected-chromedriver==3.5.5
pyarrow==15.0.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
flask-compress==1.14
# PostgreSQL dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
# Process management
psutil==5.9.5
# Celery and dependencies
celery==5.3.6
redis==5.0.1
# Validation
pydantic==2.5.2